integration_2022_wk51

See merge request oai/openairinterface5g!1877

!1823 NR UE improve RA procedures (continuation)
!1863 fixes for mixed slot without uplink symbols
!1865 Remove max power limitation in p_NR_FR1
!1866 Small fixes from Fujitsu branch
fix OAIUE test
!1638 Implementation of RRC NR_Paging message at gNB
!1846 NR limit dl_DataToUL_ACK
!1864 Derive K_NG-RAN* key for RRCReestablishment
!1871 Improvements in 5G RAR and DCI for initial BWP larger than 180PRBs
!1872 Fix seed generation and physim script
!1860 Test reduced iterations for ULSIM 3GPP
!1873 Moved nr_ue_design.md to doc/
!1878 Bugfix wrong use of qsort in dl/ul schedulers
!1881 Fix documentation of dot_product()
!1815 NR UE moved PDSCH, rxdataF to stack
!1875 NR SA SIB1 status after failure
!1870 NR SA Tutorials

+++ openair2/LAYER2/NR_MAC_COMMON/nr_mac_common.h
@@ -32,10 +32,8 @@
 #define __LAYER2_NR_MAC_COMMON_H__
 
 #include "NR_MIB.h"
-#include "NR_PDSCH-Config.h"
 #include "NR_CellGroupConfig.h"
 #include "nr_mac.h"
-#include "openair1/PHY/impl_defs_nr.h"
 #include "common/utils/nr/nr_common.h"
 
 typedef enum {
@@ -68,14 +66,14 @@ uint8_t compute_srs_resource_indicator(NR_PUSCH_ServingCellConfig_t *pusch_servi
                                        NR_PUSCH_Config_t *pusch_Config,
                                        NR_SRS_Config_t *srs_config,
                                        nr_srs_feedback_t *srs_feedback,
-                                       uint16_t *val);
+                                       uint32_t *val);
 
 uint8_t compute_precoding_information(NR_PUSCH_Config_t *pusch_Config,
                                       NR_SRS_Config_t *srs_config,
                                       dci_field_t srs_resource_indicator,
                                       nr_srs_feedback_t *srs_feedback,
                                       const uint8_t *nrOfLayers,
-                                      uint16_t *val);
+                                      uint32_t *val);
 
 uint16_t nr_dci_size(const NR_BWP_DownlinkCommon_t *initialDLBWP,
                      const NR_BWP_UplinkCommon_t *initialULBWP,
